Refactored to split into separate modules
Extracted Messages, Models, Subscriptions, Update and View into their own
modules to make working with them easier as they grow.

Also wanted to see how easy it is to do that in Elm. Even though this
was a small refactor (I'm still at the beginning with Elm) it gave me
a taste of how nice of an experience such refactors are.

Again, compiler guides you by hand with that, helping you at every step
of your way.