Removed Status
This is huge! No more Status type and status property on the Model.
Trying to keep the state as minimal as possible, deriving as much as possible
from it.

Status was used to manage the state of the calc, e.g. if we expect the operand,
which one if that's the case, or maybe we expect the operation, or maybe the
result? Turns out all that can be derived by analyzing the properties that
are or are not set on the state.