Concatenating operands
This is huge! Both, figuratively and literally as the user now can operate on
some big numbers! We also got work with the `String` and the `Result` parts
of the Elm core libraries.

`Result` is something new (to me at least). Whenever you try to perform an
operation and its outcome isn't guaranteed, you should resort to using `Result`.
It will produce either an `Ok` if everything goes well or `Err` if something
wrong happens. Then you can extract the values from it using
`case <YourResult> of`.