Added handling for decimals
This feature brought some changes to the model. It turned out that it's better
to have the operands stored as Strings, since majority of the operations
performed on them are concatenations of the user inputs. Then when calculating
the result they're being converted to Floats using String.toFloat from Elm core
String library.